Adding iree_hal_vulkan_native_buffer_t. This will allow us to import/export non-VMA buffers and support direct allocations from the Vulkan API without VMA in the middle. The VMA implementation has been cleaned up a bit to prepare it for future ejection.
