+++ runtime/src/iree/hal/drivers/vulkan/BUILD.bazel
@@ -18,6 +18,8 @@ iree_runtime_cc_library(
     name = "vulkan",
     srcs = [
         "api.cc",
+        "base_buffer.c",
+        "base_buffer.h",
         "builtin_executables.cc",
         "builtin_executables.h",
         "command_queue.h",
@@ -34,8 +36,8 @@ iree_runtime_cc_library(
         "extensibility_util.cc",
         "extensibility_util.h",
         "handle_util.h",
-        "internal_vk_mem_alloc.cc",
-        "internal_vk_mem_alloc.h",
+        "native_buffer.cc",
+        "native_buffer.h",
         "native_event.cc",
         "native_event.h",
         "native_executable.cc",
@@ -52,8 +54,8 @@ iree_runtime_cc_library(
         "tracing.h",
         "vma_allocator.cc",
         "vma_allocator.h",
-        "vma_buffer.cc",
-        "vma_buffer.h",
+        "vma_impl.cc",
+        "vma_impl.h",
         "vulkan_device.cc",
         "vulkan_driver.cc",
         "vulkan_headers.h",
