+++ runtime/src/iree/hal/drivers/vulkan/CMakeLists.txt
@@ -19,6 +19,8 @@ iree_cc_library(
     "vulkan_driver.h"
   SRCS
     "api.cc"
+    "base_buffer.c"
+    "base_buffer.h"
     "builtin_executables.cc"
     "builtin_executables.h"
     "command_queue.h"
@@ -35,8 +37,8 @@ iree_cc_library(
     "extensibility_util.cc"
     "extensibility_util.h"
     "handle_util.h"
-    "internal_vk_mem_alloc.cc"
-    "internal_vk_mem_alloc.h"
+    "native_buffer.cc"
+    "native_buffer.h"
     "native_event.cc"
     "native_event.h"
     "native_executable.cc"
@@ -53,8 +55,8 @@ iree_cc_library(
     "tracing.h"
     "vma_allocator.cc"
     "vma_allocator.h"
-    "vma_buffer.cc"
-    "vma_buffer.h"
+    "vma_impl.cc"
+    "vma_impl.h"
     "vulkan_device.cc"
     "vulkan_driver.cc"
     "vulkan_headers.h"
