+++ runtime/src/iree/hal/drivers/vulkan/descriptor_set_arena.cc
@@ -11,10 +11,10 @@
 #include <utility>
 
 #include "iree/base/internal/math.h"
+#include "iree/hal/drivers/vulkan/base_buffer.h"
 #include "iree/hal/drivers/vulkan/extensibility_util.h"
 #include "iree/hal/drivers/vulkan/native_pipeline_layout.h"
 #include "iree/hal/drivers/vulkan/status_util.h"
-#include "iree/hal/drivers/vulkan/vma_buffer.h"
 
 namespace iree {
 namespace hal {
@@ -36,10 +36,9 @@ static void PopulateDescriptorSetWriteInfos(
     const auto& binding = bindings[i];
 
     auto& buffer_info = buffer_infos[i];
-    buffer_info.buffer =
-        binding.buffer ? iree_hal_vulkan_vma_buffer_handle(
-                             iree_hal_buffer_allocated_buffer(binding.buffer))
-                       : VK_NULL_HANDLE;
+    buffer_info.buffer = binding.buffer
+                             ? iree_hal_vulkan_buffer_handle(binding.buffer)
+                             : VK_NULL_HANDLE;
     buffer_info.offset =
         iree_hal_buffer_byte_offset(binding.buffer) + binding.offset;
     if (binding.length == IREE_WHOLE_BUFFER) {
